requirements.txt
Django==1.9.5
djangorestframework==3.3.3
requests==2.2.1
pyquery==1.2.13
pymongo==3.2.2
djangotoolbox==1.8.0
django-mongodb-engine==0.6.0
